Work in Cass.py also modified the default value of keyspacePrefix in SetUp.py
Most of this work is for the new controller for accessing the IntCode
keyspace.

One note: Cassandra is weird about case, so it is best to use all
lowercase keyspace names (you might notice that you keyspaces are lower
case anyway). So, I changed the defualt value of keyspacePrefix to
'connector' from 'Connector'.